+++ gradle.properties
@@ -15,7 +15,7 @@ android.useAndroidX=true
 org.gradle.jvmargs=-Xmx2g
 
 GROUP=co.touchlab.faktory
-VERSION_NAME=0.2.6.6-SNAPSHOT
+VERSION_NAME=0.2.6.7-SNAPSHOT
 KOTLIN_VERSION=1.6.21
 
 POM_URL=https://github.com/touchlab/KMMBridge
