+++ kmmbridge/src/main/kotlin/artifactmanager/GradlePublishArtifactManager.kt
@@ -1,12 +1,23 @@
 package co.touchlab.faktory.artifactmanager
 
 import co.touchlab.faktory.publishingExtension
+import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
+import org.gradle.api.attributes.Bundling
+import org.gradle.api.attributes.Category
+import org.gradle.api.attributes.LibraryElements
+import org.gradle.api.attributes.Usage
+import org.gradle.api.attributes.java.TargetJvmVersion
+import org.gradle.api.component.SoftwareComponentFactory
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
+import org.gradle.kotlin.dsl.creating
 import org.gradle.kotlin.dsl.getByType
+import org.gradle.kotlin.dsl.getValue
+import org.gradle.kotlin.dsl.named
 import java.io.File
 
 class GradlePublishArtifactManager(
@@ -22,7 +33,16 @@ class GradlePublishArtifactManager(
     private val name: String = project.name
     private val artifactBasePath: String = "{{url}}/$group/$name/{{version}}/$name-{{version}}-kmmbridge.zip"
 
-    override fun configure(project: Project, version: String, uploadTask: Task) {
+    override fun configure(
+        project: Project,
+        version: String,
+        uploadTask: Task,
+        softwareComponentFactory: SoftwareComponentFactory
+    ) {
+        with(softwareComponentFactory.adhoc("kmmbridge")) {
+            project.components.add(this)
+            addVariantsFromConfiguration(project.createOutgoingConfiguration()) { mapToMavenScope("runtime") }
+        }
         project.publishingExtension.publications.create("SharedFramework", MavenPublication::class.java) {
             from(project.components.getByName("kmmbridge"))
             this.version = version
@@ -68,5 +88,21 @@ class GradlePublishArtifactManager(
         val repositoryName = mavenArtifactRepository.name.capitalize()
 
         return "publish${publicationName}PublicationTo${repositoryName}Repository"
+    }
+
+    private fun Project.createOutgoingConfiguration(): Configuration {
+        val configuration by configurations.creating {
+            isCanBeConsumed = true
+            isCanBeResolved = false
+            attributes {
+                attribute(Category.CATEGORY_ATTRIBUTE, objects.named(Category.LIBRARY))
+                attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage.NATIVE_LINK))
+                attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.EMBEDDED))
+                attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, JavaVersion.current().majorVersion.toInt())
+                attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named("shared-xcframework"))
+            }
+        }
+
+        return configuration
     }
 }
